Dragging
- fixes #2
- cleanups. react components are now in `components/` and common utils moved to `lib/`
- adds `react-draggable@2.2.1` dependency
