+++ lib/send-to-addon.js
@@ -1,0 +1,5 @@
+module.exports = sendToAddon;
+
+function sendToAddon(obj) {
+  window.dispatchEvent(new CustomEvent('message', {detail: obj}));
+}
