+++ package.json
@@ -41,6 +41,7 @@
     "jpm": "1.1.1",
     "react": "15.1.0",
     "react-dom": "15.1.0",
+    "react-draggable": "2.2.1",
     "watchify": "3.7.0"
   }
 }
